// src/app/shared/models/view-port-sizes.model.ts
export interface ViewPortSizes {
    isXSmall: boolean;
    isSmall: boolean;
    isMedium: boolean;
    isLarge: boolean;
    isXLarge: boolean;
    isWeb: boolean;
    isWebLandscape: boolean;
    isWebPortrait: boolean;
    isHandset: boolean;
    isHandsetLandscape: boolean;
    isHandsetPortrait: boolean;
    isTablet: boolean;
    isTabletLandscape: boolean;
    isTabletPortrait: boolean;
}